Revert "did the nested replies, to a single comment"
