+++ src/app/components/ReplySection.tsx
@@ -1,7 +1,13 @@
+"use client";
+
+import { Comment, User } from "@prisma/client";
 import React, { useState } from "react";
-import { deleteCommentById } from "../server/comment-queries/queries";
+import {
+  createComment,
+  deleteCommentById,
+} from "../server/comment-queries/queries";
 import { useSession } from "next-auth/react";
-import { Comment, Post, User } from "@prisma/client";
+import Button from "./Button";
 
 type FullComment = Comment & { user: User; replies: Comment[] };
 
@@ -11,9 +17,28 @@ type ReplySectionProps = {
 };
 
 const ReplySection = ({ comment, postId }: ReplySectionProps) => {
-  const [showReplySection, SetShowReplySection] = useState(false);
+  const [showReplySection, setShowReplySection] = useState(false);
+  const [reply, setReply] = useState("");
 
   const { data: session } = useSession();
+
+  const canDelete = comment.userEmail === session?.user?.email;
+
+  const handleReply = async (e: React.FormEvent, parnetId: number) => {
+    e.preventDefault();
+    try {
+      await createComment({
+        content: reply,
+        postId: postId,
+        userEmail: session?.user?.email!,
+        parentId: parnetId,
+      });
+      setReply("");
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
   const handleDeleteComment = async (commentId: number) => {
     try {
       await deleteCommentById({
@@ -24,35 +49,51 @@ const ReplySection = ({ comment, postId }: ReplySectionProps) => {
       console.log(error);
     }
   };
-
   return (
-    <>
-      {!!comment.replies?.length &&
-        comment.replies?.map((reply) => (
-          <div
-            className="flex flex-col rounded-md bg-gray-600 px-8 text-red-400"
-            key={reply.id}
-          >
-            <p className="  text-red-400">{reply?.content}</p>
-            <div className="space-x-3 text-right">
-              <button onClick={() => SetShowReplySection((prev) => !prev)}>
-                <p className="text-right text-xs text-gray-400">Reply</p>
-              </button>
-              <button>
-                <p
-                  onClick={() => handleDeleteComment(reply.id)}
-                  className="text-right text-xs text-gray-400"
-                >
-                  Delete
-                </p>
-              </button>
+    <div key={comment.id} className="flex items-center space-x-2">
+      <img className="size-8 rounded-full" src={comment.user.picture} />
+      <div className="flex h-fit max-w-96 flex-col rounded-lg bg-gray-200 p-2">
+        <p className="text-sm font-semibold">{comment.user?.firstname}</p>
+        <p className="max-w-fit break-words" key={comment?.id!}>
+          {comment?.content}
+        </p>
+        {/* {comment.replies?.map((reply) => {
+          return (
+            <div key={reply.id}>
+              <p className="text-red-400">{reply?.content}</p>
+              {!!reply.replies?.length && (
+                <ReplySection postId={postId} comment={reply} />
+              )}
             </div>
-            {!!reply.replies?.length && (
-              <ReplySection postId={postId} comment={reply} />
-            )}
-          </div>
-        ))}
-    </>
+          );
+        })} */}
+        <div className="flex justify-end space-x-2">
+          <button onClick={() => setShowReplySection((prev) => !prev)}>
+            <p className="text-xs text-gray-400">Reply</p>
+          </button>
+          {canDelete && (
+            <button>
+              <p
+                onClick={() => handleDeleteComment(comment.id)}
+                className="text-right text-xs text-gray-400"
+              >
+                Delete
+              </p>
+            </button>
+          )}
+        </div>
+        {showReplySection && (
+          <form onSubmit={(e) => handleReply(e, comment.id)}>
+            <input
+              value={reply}
+              type="text"
+              onChange={(e) => setReply(e.target.value)}
+            />
+            <Button text="Reply" type="submit" />
+          </form>
+        )}
+      </div>
+    </div>
   );
 };
